+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.5]
+
+### Added
+
+- Methods for resilient performance tracking.
+
+## [1.7.4]
+
+### Added
+
+- Added an environment variable to customize the SameSite attribute of the authorization token cookie
+
 ## [1.7.3]
 
 ### Fixed